Refactored lua files to use globals instead of imports and split Lobby into state management and UI (#14)
